requirements.txt
tensorflow
numpy
matplotlib
opencv-python
pillow
scikit-learn
streamlit